+++ src/scalecast/util.py
@@ -1,8 +1,39 @@
-from scalecast.Forecaster import Forecaster
+from scalecast.Forecaster import descriptive_assert, Forecaster
+import pandas_datareader as pdr
 import matplotlib.pyplot as plt
 import seaborn as sns
 import numpy as np
 
+def pdr_load(sym,start=None,end=None,src='fred',require_future_dates=True,**kwargs):
+    """ gets data using `pdr.DataReader()`
+
+    Args:
+        sym (str): the name of the series to extract.
+            cannot be an array of symbols.
+        start (str or datetime): the start date to extract data.
+        end (str or datetime): the end date to extract data.
+        src (str): the source of the API pull.
+            supported values: 'fred', 'yahoo', 'alphavantage', 'enigma', 
+            'famafrench','moex', 'quandl', 'stooq', 'tiingo'.
+        require_future_dates (bool): default True.
+            if False, none of the models from the resulting Forecaster object 
+            will forecast into future periods by default.
+            if True, all models will forecast into future periods, 
+            unless run with test_only = True, and when adding regressors, they will automatically
+            be added into future periods.
+        **kwargs passed to pdr.DataReader() function. 
+            see https://pandas-datareader.readthedocs.io/en/latest/remote_data.html.
+
+    Returns:
+        (Forecaster): a Forecaster object with the dates and y-values loaded.
+    """
+    descriptive_assert(
+        isinstance(sym,str),
+        ValueError,
+        f'sym argument only accepts str types, got {type(sym)}'
+    )
+    df = pdr.DataReader(sym,data_source=src,start=start,end=end,**kwargs)
+    return Forecaster(y=df[sym],current_dates=df.index,require_future_dates=require_future_dates)
 
 def plot_reduction_errors(f):
     """ plots the resulting error/accuracy of a Forecaster object where `reduce_Xvars()` method has been called